+++ libs/eventually/src/__tests__/utils.spec.ts
@@ -1,4 +1,4 @@
-import { camelize, decamelize } from "../utils";
+import { camelize, decamelize, formatTime } from "../utils";
 
 describe("utils", () => {
   it("should camelize", () => {
@@ -11,5 +11,14 @@ describe("utils", () => {
     expect(decamelize("ABC")).toBe("abc");
     expect(decamelize("AaaBbbCcc")).toBe("aaa-bbb-ccc");
     expect(decamelize("AaBbCc")).toBe("aa-bb-cc");
+  });
+
+  it("should format time", () => {
+    const et1 = formatTime(process.uptime());
+    expect(et1.length).toBe(5);
+    const et2 = formatTime(process.uptime() + 2 * 60 * 60);
+    expect(et2.length).toBe(8);
+    const et3 = formatTime(process.uptime() + 25 * 60 * 60);
+    expect(et3.length).toBeGreaterThan(8);
   });
 });
